+++ did_core/did_parser_nom/src/did/parsing/did_core.rs
@@ -1,17 +1,14 @@
 use nom::{
     branch::alt,
-    bytes::complete::{tag, take_till, take_while1},
-    character::complete::{alphanumeric1, char, one_of, satisfy},
-    combinator::{all_consuming, cut, fail, map, opt, recognize},
-    multi::{many1, many_m_n},
+    bytes::complete::{tag, take_while1},
+    character::complete::{alphanumeric1, char, satisfy},
+    combinator::{all_consuming, cut, opt, recognize},
+    multi::many1,
     sequence::{delimited, tuple},
     AsChar, IResult,
 };
 
-use crate::{Did, DidRange, ParseError};
-
-type DidPart<'a> = (&'a str, &'a str, Option<&'a str>, &'a str);
-pub type DidRanges = (Option<DidRange>, Option<DidRange>, Option<DidRange>);
+use super::{DidPart, DidRanges};
 
 fn hexadecimal_digit(input: &str) -> IResult<&str, &str> {
     fn is_hexadecimal_digit(c: char) -> bool {
@@ -19,12 +16,6 @@ fn hexadecimal_digit(input: &str) -> IResult<&str, &str> {
     }
 
     recognize(satisfy(is_hexadecimal_digit))(input)
-}
-
-fn base58char(input: &str) -> IResult<&str, &str> {
-    recognize(one_of(
-        "123456789ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz",
-    ))(input)
 }
 
 fn is_lowercase_alphanumeric(c: char) -> bool {
@@ -53,7 +44,7 @@ fn method_specific_id(input: &str) -> IResult<&str, &str> {
 }
 
 // namespace =  *idchar ":"
-fn namespace(input: &str) -> IResult<&str, &str> {
+pub(super) fn namespace(input: &str) -> IResult<&str, &str> {
     if let Some((before_last_colon, after_last_colon)) = input.rsplit_once(':') {
         match cut(all_consuming(many1(alt((idchar, tag(":"))))))(before_last_colon) {
             Ok(_) => Ok((after_last_colon, before_last_colon)),
@@ -68,48 +59,15 @@ fn namespace(input: &str) -> IResult<&str, &str> {
 }
 
 // did = "did:" method-name ":" method-specific-id
-fn parse_qualified_did(input: &str) -> IResult<&str, DidPart> {
+pub(super) fn parse_qualified_did(input: &str) -> IResult<&str, DidPart> {
     tuple((tag("did"), method_name, opt(namespace), method_specific_id))(input)
 }
 
-fn parse_did_web(input: &str) -> IResult<&str, DidPart> {
-    fn did_web_method(input: &str) -> IResult<&str, &str> {
-        delimited(char(':'), tag("web"), char(':'))(input)
-    }
-
-    tuple((
-        tag("did"),
-        did_web_method,
-        opt(fail::<_, &str, _>),
-        take_till(|c| "?/#".contains(c)),
-    ))(input)
-}
-
-// idstring = 21*22(base58char)
-fn parse_unqualified_sovrin_did(input: &str) -> IResult<&str, &str> {
-    recognize(many_m_n(21, 22, base58char))(input)
-}
-
-// The specification seems to contradict practice?
-// sovrin-did = "did:sov:" idstring *(":" subnamespace)
-// subnamespace = ALPHA *(ALPHA / DIGIT / "_" / "-")
-fn parse_qualified_sovrin_did(input: &str) -> IResult<&str, DidPart> {
-    fn did_sov_method(input: &str) -> IResult<&str, &str> {
-        delimited(char(':'), tag("sov"), char(':'))(input)
-    }
-    tuple((
-        tag("did"),
-        did_sov_method,
-        opt(namespace),
-        cut(parse_unqualified_sovrin_did),
-    ))(input)
-}
-
-fn to_id_range(id: &str) -> DidRanges {
+pub(super) fn to_id_range(id: &str) -> DidRanges {
     (None, None, Some(0..id.len()))
 }
 
-fn to_did_ranges((did_prefix, method, namespace, id): DidPart) -> DidRanges {
+pub(super) fn to_did_ranges((did_prefix, method, namespace, id): DidPart) -> DidRanges {
     let mut next_start = if !did_prefix.is_empty() {
         did_prefix.len() + 1
     } else {
@@ -141,33 +99,4 @@ fn to_did_ranges((did_prefix, method, namespace, id): DidPart) -> DidRanges {
     };
 
     (method_range, namespace_range, id_range)
-}
-
-pub fn parse_did_ranges(input: &str) -> IResult<&str, DidRanges> {
-    alt((
-        map(parse_did_web, to_did_ranges),
-        map(parse_qualified_sovrin_did, to_did_ranges),
-        map(parse_qualified_did, to_did_ranges),
-        map(parse_unqualified_sovrin_did, to_id_range),
-    ))(input)
-}
-
-pub fn parse_did(did: String) -> Result<Did, ParseError> {
-    if did.is_empty() {
-        return Err(ParseError::InvalidInput("Empty input"));
-    }
-
-    let (_, (method, namespace, id)) = all_consuming(parse_did_ranges)(&did)?;
-    let id = id.ok_or_else(|| ParseError::InvalidInput("Invalid DID"))?;
-
-    if id.end > did.len() {
-        return Err(ParseError::InvalidInput("Invalid DID"));
-    }
-
-    Ok(Did {
-        did,
-        method,
-        namespace,
-        id,
-    })
 }
